+++ resources/views/livewire/lead/activities/meetings/show-meeting.blade.php
@@ -1,32 +1,37 @@
 <div>
-    <x-activity-layout title="Show Call ({{ $activity->title }})" :activity="$activity" :notes="$notes">
+    <x-activity-layout title="Show Meeting ({{ $activity->title }})" :activity="$activity" :notes="$notes">
         <x-slot:header>
             <li class="breadcrumb-item"><a href="{{ route('leads.show', ['lead' => $activity->lead->id]) }}"
                     wire:navigate>{{ $activity->lead->name }}</a></li>
         </x-slot:header>
         <x-slot:tabsHeader>
             <li class="nav-item" role="presentation">
-                <button class="nav-link border active" id="call-tab-tab" data-bs-toggle="pill"
-                    data-bs-target="#call-tab-pane" type="button" role="tab" aria-controls="call-tab-pane"
-                    aria-selected="true" tabindex="-1">Call</button>
+                <button class="nav-link border active" id="meeting-tab-tab" data-bs-toggle="pill"
+                    data-bs-target="#meeting-tab-pane" type="button" role="tab" aria-controls="meeting-tab-pane"
+                    aria-selected="true" tabindex="-1">Meeting</button>
             </li>
         </x-slot:tabsHeader>
         <x-slot:tabsContent>
-            <div class="tab-pane fade p-0 border-0 show active" id="call-tab-pane" role="tabpanel"
+            <div class="tab-pane fade p-0 border-0 show active" id="meeting-tab-pane" role="tabpanel"
                 aria-labelledby="posts-tab-pane" tabindex="0">
                 <div class="d-flex justify-content-between">
                     <h6 class="fw-semibold">Details</h6>
                     <div>
-                        <a class="btn btn-outline-warning btn-wave btn"
-                            href="{{ route('leads.activities.calls.edit', ['lead' => $activity->lead_id, 'call' => $activity->id]) }}"
-                            wire:navigate>
-                            <i class="ti ti-pencil"></i>
-                        </a>
-                        <button type="button" class="btn btn-outline-danger btn-wave  btn" data-bs-toggle="modal"
-                            data-bs-target="#DeleteActivityModal"
-                            x-on:click="id = {{ $activity->id }}; title = '{{ $activity->title }}';">
-                            <i class="ti ti-trash"></i>
-                        </button>
+                        @can(\App\Enums\PermissionEnum::CRM_ACTIVIY_OPERATION->value)
+                            <a class="btn btn-outline-warning btn-wave btn"
+                                href="{{ route('leads.activities.meetings.edit', ['lead' => $activity->lead_id, 'meeting' => $activity->id]) }}"
+                                wire:navigate>
+                                <i class="ti ti-pencil"></i>
+                            </a>
+                        @endcan
+
+                        @can(\App\Enums\PermissionEnum::CRM_ACTIVIY_DELETE->value)
+                            <button type="button" class="btn btn-outline-danger btn-wave  btn" data-bs-toggle="modal"
+                                data-bs-target="#DeleteActivityModal"
+                                x-on:click="id = {{ $activity->id }}; title = '{{ $activity->title }}';">
+                                <i class="ti ti-trash"></i>
+                            </button>
+                        @endcan
                     </div>
                 </div>
                 <div class="table-responsive mb-3 text-muted">
@@ -81,65 +86,68 @@
                             <tr class="product-list">
                                 <td class="text-muted">
                                     <div class="fw-semibold">
-                                        Call Type
+                                        Meeting Type
                                     </div>
                                 </td>
                                 <td class="text-muted">
                                     :
                                 </td>
-                                <td class="text-muted">{{ $activity->activityable->type }}
+                                <td class="text-muted">{{ $activity->activityable->online ? 'Online' : 'Offline' }}
                                 </td>
                             </tr>
                             <tr class="product-list">
                                 <td class="text-muted">
                                     <div class="fw-semibold">
-                                        Date
+                                        Start Date
                                     </div>
                                 </td>
                                 <td class="text-muted">
                                     :
                                 </td>
                                 <td class="text-muted">
-                                    {{ $activity->activityable->call_date }}
+                                    {{ $activity->activityable->start }}
                                 </td>
                             </tr>
                             <tr class="product-list">
                                 <td class="text-muted">
                                     <div class="fw-semibold">
-                                        Duration Call
+                                        End Date
                                     </div>
                                 </td>
                                 <td class="text-muted">
                                     :
                                 </td>
                                 <td class="text-muted">
-                                    {{ $activity->activityable->duration_call }}
+                                    {{ $activity->activityable->end }}
                                 </td>
                             </tr>
                             <tr class="product-list">
                                 <td class="text-muted">
                                     <div class="fw-semibold">
-                                        Call Reason
+                                        Link
                                     </div>
                                 </td>
                                 <td class="text-muted">
                                     :
                                 </td>
                                 <td class="text-muted">
-                                    {{ $activity->activityable->callReason->name ?? 'N/A' }}
+                                    <a href="{{ $activity->activityable->link ?? '#' }}" target="_blank"
+                                        class="text-primary" rel="noopener noreferrer">
+                                        {{ str($activity->activityable->link)->limit(30) ?? 'N/A' }}
+                                    </a>
                                 </td>
                             </tr>
                             <tr class="product-list">
                                 <td class="text-muted">
                                     <div class="fw-semibold">
-                                        Call Response
+                                        Location
                                     </div>
                                 </td>
                                 <td class="text-muted">
                                     :
                                 </td>
                                 <td class="text-muted">
-                                    {{ $activity->activityable->callResponse->name ?? 'N/A' }}
+                                    {{ $activity->activityable->location ?? 'N/A' }}
                                 </td>
                             </tr>
                             <tr class="product-list">
@@ -162,7 +170,7 @@
             <div class="card custom-card">
                 <div class="card-header justify-content-between">
                     <div class="card-title">
-                        Recent Calls
+                        Recent Meetings
                     </div>
                     <div class="">
                         <a href="javascript:void(0);" class="btn btn-outline-light btn-sm">
@@ -173,7 +181,7 @@
                 <div class="card-body">
                     <div>
                         <ul class="list-unstyled mb-0 crm-recent-activity">
-                            @foreach ($calls as $item)
+                            @foreach ($meetings as $item)
                                 <li class="crm-recent-activity-content">
                                     <div class="d-flex align-items-top">
                                         <div class="me-3">
@@ -183,13 +191,13 @@
                                         </div>
                                         <div class="crm-timeline-content">
                                             <div>
-                                                <span class="fw-semibold">New Call &amp;</span><span><a
-                                                        href="{{ route('leads.activities.calls.edit', ['lead' => $item->lead->id, 'call' => $item->id]) }}"
+                                                <span class="fw-semibold">New Meeting &amp;</span><span><a
+                                                        href="{{ route('leads.activities.meetings.edit', ['lead' => $item->lead->id, 'meeting' => $item->id]) }}"
                                                         class="text-primary fw-semibold">
                                                         {{ $item->title }}.</a></span>
                                             </div>
                                             <p>
-                                                {{ $item->activityable->call_date }}
+                                                {{ $item->activityable->start }}
                                             </p>
                                         </div>
                                         <div class="flex-fill text-end">
@@ -204,24 +212,27 @@
                 </div>
             </div>
         </x-slot:sidebar>
-        <x-slot:modals>
-            <!-- Start::DeleteNoteModal -->
-            <x-modal id="DeleteNoteModal">
-                <x-slot:title>
-                    <i class="ti ti-trash text-danger me-1"></i>
-                    <span>
-                        Are you sure you want to delete this Note?!
-                    </span>
-                </x-slot:title>
-                <x-slot:content>
-                    <p>Are you sure you want to delete this note</p>
-                </x-slot:content>
-                <x-slot:footer>
-                    <button type="button" class="btn btn-success" data-bs-dismiss="modal">Close</button>
-                    <button type="button" class="btn btn-danger" wire:click="deleteNote(noteId)">Delete</button>
-                </x-slot:footer>
-            </x-modal>
-            <!-- End::DeleteNoteModal -->
-        </x-slot:modals>
+
+        @can(\App\Enums\PermissionEnum::CRM_ACTIVIY_DELETE->value)
+            <x-slot:modals>
+                <!-- Start::DeleteNoteModal -->
+                <x-modal id="DeleteNoteModal">
+                    <x-slot:title>
+                        <i class="ti ti-trash text-danger me-1"></i>
+                        <span>
+                            Are you sure you want to delete this Note?!
+                        </span>
+                    </x-slot:title>
+                    <x-slot:content>
+                        <p>Are you sure you want to delete this note</p>
+                    </x-slot:content>
+                    <x-slot:footer>
+                        <button type="button" class="btn btn-success" data-bs-dismiss="modal">Close</button>
+                        <button type="button" class="btn btn-danger" wire:click="deleteNote(noteId)">Delete</button>
+                    </x-slot:footer>
+                </x-modal>
+                <!-- End::DeleteNoteModal -->
+            </x-slot:modals>
+        @endcan
     </x-activity-layout>
 </div>
